requirements.txt
mediapipe>=0.10.0
opencv-python
numpy

# Flask Backend (Phase 2.0/2.1)
Flask>=3.0.0
Flask-CORS>=4.0.0
python-dotenv>=1.0.0
werkzeug>=3.0.0  # For password hashing (Phase 2.1)

# MongoDB Database (Phase B)
pymongo>=4.6.0  # MongoDB driver 